+++ src/trufo/crypto/awskms/__init__.py
@@ -1,6 +1,0 @@
-# Copyright 2025-2026 Trufo, Inc. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""AWS KMS signing primitives."""
-
-from .awskms_ecsk import EllipticCurvePrivateKeyAwsKms